♻️ Refactor filter with react query
