+++ src/pages/problem/long/LongProblemListPage.tsx
@@ -3,21 +3,31 @@ import { longProblemApiWrapper } from '../../../api/wrapper/problem/longProblemA
 import { useFilter } from '../../../hooks/useFilter';
 import { useTable } from '../../../hooks/useTable';
 import { useQuery } from 'react-query';
-import { IProblemListData } from '../../../types/problem/api';
+import { IProblemListData, IProblemListRequest } from '../../../types/problem/api';
 import { ProblemListPageTemplate } from '../ProblemListPageTemplate';
 import { longProblemTableHeads } from '../../../constants/tableHeads';
+import { useEffect, useState } from 'react';
 
 export const LongProblemListPage = () => {
+  const [params, setParams] = useState<IProblemListRequest>();
   const filterHandler = useFilter();
-  const { page, handleChangePage, handleRowClick, params } = useTable(
-    filterHandler.filterState,
-    URLWithParam.LONG_PROBLEM_DETAIL,
-  );
+  const { page, handleChangePage, handleRowClick } = useTable(URLWithParam.LONG_PROBLEM_DETAIL);
   const { data } = useQuery<IProblemListData>(
     ['longProblemList', params],
     () => longProblemApiWrapper.getLongProblemList(params),
     { enabled: !!params },
   );
+
+  useEffect(() => {
+    setParams({ ...filterHandler.getParams(), page: page });
+  }, [page]);
+
+  useEffect(() => {
+    if (filterHandler.filterState.length > 0 && filterHandler.filterState.some((e) => e.value))
+      setParams((prev) => {
+        return { ...prev, ...filterHandler.getParams() };
+      });
+  }, [filterHandler.filterState]);
 
   return (
     <ProblemListPageTemplate
